+++ CHANGELOG.md
@@ -5,6 +5,8 @@
       `ZoneSorterByOffsetAndName`.
         * AVR: 180-530 bytes of flash
         * 32-bit: 120-600 bytes of flash
+    * Rename internal `TransitionStorage::findTransition()` to
+      `findTransitionForSeconds()` for clarify and self-documenation.
 * 1.9.0 (2021-12-02, TZDB 2021e)
     * Add `ZoneSorterByName` and `ZoneSorterByOffsetAndName` classes
       to sort zone indexes, ids, or names according to 2 pre-defined sorting